+++ apps/community/Leads/Models/LeadService.php
@@ -59,8 +59,9 @@ class LeadService extends \HubletoMain\Core\Model
   public function tableDescribe(array $description = []): array
   {
     $description = parent::tableDescribe();
-    if ($this->main->urlParamAsInteger('idLead') > 0){
+    if ($this->main->urlParamAsInteger('idLead') > 0) {
       // custom table description
     }
+    return $description;
   }
 }